+++ amazon-chime-sdk/src/main/java/com/amazonaws/services/chime/sdk/meetings/internal/video/DefaultVideoClientController.kt
@@ -6,7 +6,6 @@
 package com.amazonaws.services.chime.sdk.meetings.internal.video
 
 import android.content.Context
-import com.amazonaws.services.chime.sdk.BuildConfig
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.video.VideoSource
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.video.capture.CameraCaptureSource
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.video.capture.DefaultCameraCaptureSource
@@ -14,6 +13,7 @@ import com.amazonaws.services.chime.sdk.meetings.audiovideo.video.capture.Defaul
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.video.gl.EglCore
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.video.gl.EglCoreFactory
 import com.amazonaws.services.chime.sdk.meetings.device.MediaDevice
+import com.amazonaws.services.chime.sdk.meetings.internal.utils.AppInfoUtil
 import com.amazonaws.services.chime.sdk.meetings.session.MeetingSessionConfiguration
 import com.amazonaws.services.chime.sdk.meetings.utils.logger.Logger
 import com.google.gson.Gson
@@ -41,6 +41,7 @@ class DefaultVideoClientController(
     private val VIDEO_CLIENT_FLAG_ENABLE_USE_HW_DECODE_AND_RENDER = 1 shl 6
     private val VIDEO_CLIENT_FLAG_ENABLE_TWO_SIMULCAST_STREAMS = 1 shl 12
     private val VIDEO_CLIENT_FLAG_DISABLE_CAPTURER = 1 shl 20
+    private val VIDEO_CLIENT_FLAG_EXCLUDE_SELF_CONTENT_IN_INDEX = 1 shl 24
 
     private val gson = Gson()
 
@@ -180,8 +181,8 @@ class DefaultVideoClientController(
 
     override fun initializeVideoClient() {
         logger.info(TAG, "Initializing video client")
-        initializeAppDetailedInfo()
-        VideoClient.initializeGlobals(context)
+        AppInfoUtil.initializeVideoClientAppDetailedInfo(context)
+        VideoClient.javaInitializeGlobals(context)
         VideoClientCapturer.getInstance(context)
         videoClient = videoClientFactory.getVideoClient(videoClientObserver)
     }
@@ -193,7 +194,8 @@ class DefaultVideoClientController(
         flag = flag or VIDEO_CLIENT_FLAG_ENABLE_USE_HW_DECODE_AND_RENDER
         flag = flag or VIDEO_CLIENT_FLAG_ENABLE_TWO_SIMULCAST_STREAMS
         flag = flag or VIDEO_CLIENT_FLAG_DISABLE_CAPTURER
-        videoClient?.startServiceV3(
+        flag = flag or VIDEO_CLIENT_FLAG_EXCLUDE_SELF_CONTENT_IN_INDEX
+        videoClient?.javaStartServiceV3(
             "",
             "",
             configuration.meetingId,
@@ -204,12 +206,12 @@ class DefaultVideoClientController(
             eglCore?.eglContext
         )
 
-        videoSourceAdapter?.let { videoClient?.setExternalVideoSource(it, eglCore?.eglContext) }
+        videoSourceAdapter.let { videoClient?.setExternalVideoSource(it, eglCore?.eglContext) }
     }
 
     override fun stopVideoClient() {
         logger.info(TAG, "Stopping video client")
-        videoClient?.stopService()
+        videoClient?.javaStopService()
     }
 
     override fun destroyVideoClient() {
@@ -217,27 +219,5 @@ class DefaultVideoClientController(
         videoClient?.destroy()
         videoClient = null
         VideoClient.finalizeGlobals()
-    }
-
-    private fun initializeAppDetailedInfo() {
-        val manufacturer = android.os.Build.MANUFACTURER
-        val model = android.os.Build.MODEL
-        val osVersion = android.os.Build.VERSION.RELEASE
-        val packageName = context.packageName
-        val packageInfo = context.packageManager.getPackageInfo(packageName, 0)
-        val appVer = packageInfo.versionName
-        val appCode = packageInfo.versionCode.toString()
-        val clientSource = "amazon-chime-sdk"
-        val sdkVersion = BuildConfig.VERSION_NAME
-
-        VideoClient.AppDetailedInfo.initialize(
-            String.format("Android %s", appVer),
-            appCode,
-            model,
-            manufacturer,
-            osVersion,
-            clientSource,
-            sdkVersion
-        )
     }
 }