+++ app/src/main/java/com/amazonaws/services/chime/sdkdemo/data/VideoCollectionTile.kt
@@ -4,11 +4,22 @@
  */
 package com.amazonaws.services.chime.sdkdemo.data
 
+import androidx.constraintlayout.widget.ConstraintLayout
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.video.DefaultVideoRenderView
 import com.amazonaws.services.chime.sdk.meetings.audiovideo.video.VideoTileState
 
 data class VideoCollectionTile(
     val attendeeName: String,
-    val videoTileState: VideoTileState,
-    var videoRenderView: DefaultVideoRenderView?
-)
+    val videoTileState: VideoTileState
+) {
+    var videoRenderView: DefaultVideoRenderView? = null
+    var pauseMessageView: ConstraintLayout? = null
+
+    fun setRenderViewVisibility(visibility: Int) {
+        videoRenderView?.visibility = visibility
+    }
+
+    fun setPauseMessageVisibility(visibility: Int) {
+        pauseMessageView?.visibility = visibility
+    }
+}